Fixed some bugs/icks related to CurrencySelectorButton.
