+++ android/app/src/main/java/com/example/budgiet/MainActivity.kt
@@ -1,8 +1,6 @@
 package com.example.budgiet
 
-import android.content.Context
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
@@ -23,23 +21,18 @@ import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.State
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import androidx.compose.runtime.toMutableStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.budgiet.ui.NewTransactionForm
 import com.example.budgiet.ui.theme.BudgietTheme
 import com.example.budgiet.ui.utils.PlainToolTipBox
-import java.io.File
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -94,116 +87,10 @@ fun MainPage(modifier: Modifier = Modifier) {
     }
 }
 
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-// TODO: semantic test these 3 functions
-
-/* FIXME: I want to move all of this to Price.kt,
- * but for some reason it has a bug that the loading indicator shows up until the recentlyUsedCurrencies list is modified.
- * This only happens when i try to put it in Price.kt, but not when it is in MainActivity.kt. */
-private val recentlyUsedCurrencies: MutableState<Result<SnapshotStateList<String>>?> = mutableStateOf(null)
-private const val RECENT_CURRENCIES_FILE_NAME = "recentCurrencies.txt"
-private const val RECENT_CURRENCIES_LOG_TAG = "RecentlyUsedCurrencies"
-
-/** Returns an ordered [List] of **currency codes**, sorted by *most recent use*.
+/** ***WARNING!***: Don't use this directly. Instead, use [getRecentlySelectedCurrencies], [clearRecentlyUsedCurrencies], or [markCurrencyRecentlyUsed].
  *
- * The return value tells the state of the data:
- *  * **`null`**: The data is still being loaded.
- *  * **[Result.Err]**: There was an error loading the data.
- *  * **[Result.Ok]**: The data finished loading successfully.
- *
- *  Since this is a [Composable] with an internal [MutableState],
- *  changes in the state will propagate to the caller and it will be recomposed,
- *  even if this function itself does not return a [MutableState]. */
-@Composable
-fun getRecentlySelectedCurrencies(): State<Result<List<String>>?> {
-    if (recentlyUsedCurrencies.value == null) {
-        // Load ordered currencies from storage.
-
-        val context = LocalContext.current
-        val result = rememberWork(recentlyUsedCurrencies) {
-            File(context.filesDir, RECENT_CURRENCIES_FILE_NAME)
-                // Read the entirety of the file to move around the elements.
-                .readText()
-                .split('\n')
-                // Last element will always be empty because the file always ends with newLine (unless it is empty).
-                .dropLast(1)
-                .toMutableStateList()
-        }
-
-        if (result.value is Result.Err) {
-            Log.e(RECENT_CURRENCIES_LOG_TAG, "Error reading recent currencies from storage: ${(result.value as Result.Err).error}")
-        }
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    return recentlyUsedCurrencies as State<Result<List<String>>?>
-}
-
-/** Removes (clears) all currencies from the *ordered list* in memory and from the file in storage. */
-fun Context.clearRecentlyUsedCurrencies() {
-    // Clear in memory
-    recentlyUsedCurrencies.value = Result.Ok(mutableStateListOf())
-    // Clear in storage
-    dispatchWork {
-        Log.i(RECENT_CURRENCIES_LOG_TAG, "Clear recently used currencies in storage.")
-        val file = File(this.filesDir, RECENT_CURRENCIES_FILE_NAME)
-        file.writeText("")
-    }
-}
-
-/** Marks a [Currency][java.util.Currency] as recently used (a.k.a. it was just selected),
- * moving it to the front of the [List] of recent currencies,
- * which is **sorted** by latest use.
- *
- * This function will also write to the [File] in storage the same content as the [List] in memory.
- *
- * See [getRecentlySelectedCurrencies] to read from this [List]. */
-fun Context.markCurrencyRecentlyUsed(currencyCode: String) {
-    when (recentlyUsedCurrencies.value) {
-        null, is Result.Err -> recentlyUsedCurrencies.value = Result.Ok(mutableStateListOf())
-        is Result.Ok -> { }
-    }
-    val orderedCurrencies = recentlyUsedCurrencies.value!!.getOkOrNull()!! as MutableList<String>
-
-    Log.i(RECENT_CURRENCIES_LOG_TAG, "Moving Currency \"$currencyCode\" to the front of MutableStateList in memory.")
-
-    // Apply to mutable list in memory
-    // Find currency in the argument
-    when (val idx = orderedCurrencies.indexOf(currencyCode)) {
-        // The currency was already first in the list; do nothing.
-        0 -> { }
-        // Currency was not found in the List, so it must be prepended.
-        -1 -> orderedCurrencies.add(0, currencyCode)
-        // Remove target currency (arg) from the List, and put it in the front.
-        else -> {
-            orderedCurrencies.add(0, orderedCurrencies.removeAt(idx))
-        }
-    }
-
-    // Apply to storage
-    dispatchWork {
-        Log.i(RECENT_CURRENCIES_LOG_TAG, "Moving Currency \"$currencyCode\" to the front of File in storage.")
-        val file = File(this.filesDir, RECENT_CURRENCIES_FILE_NAME)
-        file.createNewFile()
-        // Write the modified list
-        file.writeText(orderedCurrencies.joinToString(separator = "", truncated = "") { code -> "$code\n" })
-    }
-}
-
+ * This should go in `Price.kt`, but putting it there prevents recomposition (for some reason) even when the MutableState is updated. */
+internal val recentlyUsedCurrencies: MutableState<Result<SnapshotStateList<String>>?> = mutableStateOf(null)
 
 class Location(
     val id: UInt,
